// Frontend/src/swagger/model/expenseDetails.ts
/**
 * Expense Tracker
 * This is the API routes for Expense Tracker
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface ExpenseDetails { 
    expenseAmount?: number;
    description?: string;
    category?: string;
    _id?: string
}